requirements/base.txt
# Core
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
django-cors-headers>=4.3,<4.4
djangorestframework-simplejwt>=5.3,<5.4
django-environ>=0.11,<0.12
django-filter>=23.5,<24.0
django-mptt>=0.16,<0.17
drf-spectacular>=0.27,<0.28

# Database
psycopg2-binary>=2.9,<3.0

# Media
Pillow>=10.2,<10.3

# Utilities
python-slugify>=8.0,<9.0

# Optional (Redis, Celery)
redis>=5.0,<6.0
celery>=5.3,<5.4
django-redis>=5.4,<5.5
